pyproject.toml
[tool.poetry]
name = "pipelines"
version = "0.1.0"
description = ""
authors = ["Gabriel Gazola Milan <gabriel.gazola@poli.ufrj.br>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
prefect = "1.4.1"
google-cloud-storage = "^2.10.0"
infisical = "^1.4.0"
basedosdados = "2.0.0b5"


[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
pre-commit = "^3.3.3"
taskipy = "^1.12.0"
isort = "^5.12.0"
semgrep = "^1.34.1"


[tool.poetry.group.ci]
optional = true

[tool.poetry.group.ci.dependencies]
networkx = "^3.1"
loguru = "^0.7.0"
typer = "^0.9.0"

[tool.black]
line-length = 100
target-version = ["py310"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "semgrep scan --error --config auto --exclude-rule yaml.github-actions.security.third-party-action-not-pinned-to-commit-sha.third-party-action-not-pinned-to-commit-sha && black . && isort . && flake8 ."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
